+++ gcs/CHANGES.md
@@ -3,6 +3,7 @@
 ## Next
 1. Skip soft delete for compose operations using deleteSourceObjects
 1. Upgrade google.api-client-libraries.version to 2.9.0
+1. Add fs.gs.universe.domain support for Google multi-universe (TPC)
 
 ## 4.0.5 - 2026-08-26